site_lions_belg/requirements.txt
aldryn-boilerplates==0.8.0
aldryn-forms==4.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
dj-database-url==0.5.0
Django==2.2.12
django-absolute==0.3
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-cms==3.7.1
django-emailit==0.2.4
django-filer==1.7.0
django-formtools==2.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-ranged-response==0.2.0
django-sekizai==1.1.0
django-simple-captcha==0.5.12
django-sizefield==1.0.0
django-tablib==3.2
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.5.0
djangocms-file==2.4.0
djangocms-forms==0.2.5
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.2.1
djangocms-light-gallery==1.1.0
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
hashids==1.2.0
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
jsonfield==3.0.0
lxml==4.5.0
mysqlclient==1.4.6
openpyxl==2.4.9
Pillow==7.1.1
pkg-resources==0.0.0
polls==0.0.4
premailer==3.6.1
pytz==2019.3
requests==2.23.0
reverse==0.1.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
transliterate==1.10.2
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.9
webencodings==0.5.1
YURL==1.0.0
